# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoteRequest": {
        "description": "POST /shortlists/{shortlist_id}/recipes/{recipe_id}/vote body.",
        "properties": {
          "vote": {
            "enum": [
              "yes",
              "no"
            ],
            "title": "Vote",
            "type": "string"
          }
        },
        "required": [
          "vote"
        ],
        "title": "VoteRequest",
        "type": "object"
      },
      "VoteResponse": {
        "description": "Mirrors the broadcast payload — frontend uses this shape unchanged\nfor both HTTP response and `vote.created` WS event handling.\n\nPhase 41 (UNDO-01): `vote_id` added — the frontend stores this on a\nsuccessful POST so it can later call DELETE /votes/{vote_id} for undo.\nPer CLAUDE.md architecture invariant #2, voting state stays computed\n(no `state` column); the undo flow is a row delete, not a state flip.",
        "properties": {
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "recipe_id": {
            "format": "uuid",
            "title": "Recipe Id",
            "type": "string"
          },
          "shortlist_id": {
            "format": "uuid",
            "title": "Shortlist Id",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "vote": {
            "title": "Vote",
            "type": "string"
          },
          "vote_id": {
            "format": "uuid",
            "title": "Vote Id",
            "type": "string"
          }
        },
        "required": [
          "vote_id",
          "shortlist_id",
          "recipe_id",
          "member_id",
          "vote",
          "state"
        ],
        "title": "VoteResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Al Dente API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/shortlists/{shortlist_id}/recipes/{recipe_id}/vote": {
      "post": {
        "description": "VOTE-01 + VOTE-02 + VOTE-05: upsert + compute state + broadcast.",
        "operationId": "cast_vote_shortlists__shortlist_id__recipes__recipe_id__vote_post",
        "parameters": [
          {
            "in": "path",
            "name": "shortlist_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shortlist Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "recipe_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recipe Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "aldente_auth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aldente Auth"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cast Vote",
        "tags": [
          "votes"
        ]
      }
    },
    "/votes/{vote_id}": {
      "delete": {
        "description": "UNDO-01: hard-delete one vote row. Architecture invariant #2 holds —\nstate is computed from row existence, so deleting the row IS the undo.\n\nOrder matters:\n  1. Resolve the vote (404 if missing — invariant #2 record-existence non-leak).\n  2. Cross-household check via Member.household_id join (Vote has no\n     household_id column — isolation runs through Member). Same 404 on\n     mismatch, NOT 403 — D-38-02 / invariant #2.\n  3. Resolve shortlist for the date we use in the veto-window guard.\n  4. Veto-window guard (D-12): if any CookingLog exists for\n     (household_id, shortlist.date) — counted by date(cooked_at) — refuse\n     with 409 `veto_window_closed`. Defense-in-depth behind the frontend's\n     preemptive disabled button (Plan 41-04).\n  5. Snapshot the broadcast payload BEFORE delete (FKs still resolvable).\n  6. Delete + commit.\n  7. Broadcast `vote.deleted`. Receiving clients drop the row from their\n     votes[] cache and `compute_vote_state` naturally re-derives.",
        "operationId": "delete_vote_votes__vote_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "vote_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vote Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "aldente_auth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aldente Auth"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Vote",
        "tags": [
          "votes"
        ]
      }
    }
  }
}